feat: add new module to compute pickup schedule
Given a schedule name (e.g. monA) and a date (e.g today) this
module will compute the next pickup day and what services are
provided for Madison, WI.

NOTE: It does not account for holidays, yet. There's also going
to be some shenanigans around timezones and UTC time for which we'll
eventually have to account.