feat: uses collectionlookup.cfm to determine schedule
Makes a form post to the city of madison to resolve the
city schedule. The response to the post is a redirect to
the schedule's pdf... we don't follow it and just read the
location header to determine its name. From the name we can
figure out the pickup schedule.